moved InstanceMethods module to its own file
